Create program that controls brightness of pixels
By time, buttons and an incremental rotary encoder attached to pins A1 (CLK) A2 (DT) and A3 (SW)

For some reason signal from encoder is reported as True unless it's activated (clicked or rotated). That's why the values are negated in code. I tried to change pull value to UP, but apparently it did not have any effect.